add batch update mode to MAB simulation (#579)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/579

Add batch training mode to the simulation. In batch mode the model is updated every N steps.

Reviewed By: czxttkl

Differential Revision: D32411860

fbshipit-source-id: f700713d443ddc1c91ffa84513a3c76771bea72a

+++ reagent/mab/mab_algorithm.py
@@ -189,7 +189,7 @@ class MABAlgo(torch.nn.Module, ABC):
         # set `inf` scores for arms which don't have the minimum number of observations
         return torch.where(
             self.total_n_obs_per_arm >= self.min_num_obs_per_arm,
-            self.get_scores(),
+            self.get_scores().float(),
             torch.tensor(torch.inf, dtype=torch.float),
         )
